+++ cl_dll/hud/ammohistory.cpp
@@ -128,7 +128,7 @@ int HistoryResource :: DrawAmmoHistory( float flTime )
 				int r, g, b;
 				DrawUtils::UnpackRGB(r,g,b, RGB_YELLOWISH);
 				float scale = (rgAmmoHistory[i].DisplayTime - flTime) * 80;
-				DrawUtils::ScaleColors(r, g, b, min(scale, 255) );
+				DrawUtils::ScaleColors(r, g, b, min((int)scale, 255) );
 
 				// Draw the pic
 				int ypos = ScreenHeight - (AMMO_PICKUP_PICK_HEIGHT + (AMMO_PICKUP_GAP * i));
@@ -156,7 +156,7 @@ int HistoryResource :: DrawAmmoHistory( float flTime )
 					DrawUtils::UnpackRGB(r,g,b, RGB_REDISH);	// if the weapon doesn't have ammo, display it as red
 
 				float scale = (rgAmmoHistory[i].DisplayTime - flTime) * 80;
-				DrawUtils::ScaleColors(r, g, b, min(scale, 255) );
+				DrawUtils::ScaleColors(r, g, b, min((int)scale, 255) );
 
 				int ypos = ScreenHeight - (AMMO_PICKUP_PICK_HEIGHT + (AMMO_PICKUP_GAP * i));
 				int xpos = ScreenWidth - (weap->rcInactive.right - weap->rcInactive.left);
@@ -174,7 +174,7 @@ int HistoryResource :: DrawAmmoHistory( float flTime )
 
 				DrawUtils::UnpackRGB(r,g,b, RGB_YELLOWISH);
 				float scale = (rgAmmoHistory[i].DisplayTime - flTime) * 80;
-				DrawUtils::ScaleColors(r, g, b, min(scale, 255) );
+				DrawUtils::ScaleColors(r, g, b, min((int)scale, 255) );
 
 				int ypos = ScreenHeight - (AMMO_PICKUP_PICK_HEIGHT + (AMMO_PICKUP_GAP * i));
 				int xpos = ScreenWidth - (rect.right - rect.left) - 10;
